[FLINK-21756][jmx] Move JMXServer to flink-core
